feat: creating the welcome screen and extracting the logic to validate the link signature
